Metrics Unit test work reviewed
1. Metrics: executorService getter is no more needed, removed;
2. MetricsTests:
  2.1 moved under /analytics package;
  2.2 countWithDoAsyncFalsePersists improved;
